VxCentralScan: Move system admin screen to separate file
